+++ src/main/java/org/example/carservice/service/MasterService.java
@@ -1,5 +1,6 @@
 package org.example.carservice.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import org.example.carservice.model.Master;
 import org.example.carservice.model.Order;
@@ -11,7 +12,7 @@ public interface MasterService {
 
     List<Order> findAllOrdersByMasterId(Long masterId);
 
-    double getSalaryOfMasterByOrder(Long masterId, Long orderId);
+    BigDecimal getSalaryOfMasterByOrder(Long masterId, Long orderId);
 
     Master getById(Long id);
